Move GameType to client-types
