+++ src/client/setup/setup-game.tsx
@@ -1,8 +1,8 @@
 import { Button, H3, H6, Slider } from "@blueprintjs/core";
 import { Dispatch, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { Difficulty } from "../../common/client-types";
-import { GameType, Side } from "../../common/game-types";
+import { Difficulty, GameType } from "../../common/client-types";
+import { Side } from "../../common/game-types";
 import { post } from "../api";
 
 enum DesiredSide {
