+++ src/common/game-types.ts
@@ -40,10 +40,6 @@ export class Piece {
     ) {}
 }
 
-export enum GameType {
-    COMPUTER = "computer",
-    HUMAN = "human",
-}
 export interface Move {
     from: Square;
     to: Square;
